UI layout
I copied the UI from the Drone scene, brought it to Olga Scene and changed it to support automatic layout (scales to screen).
I made a prefab out of it, to transport it throughout scenes.
You can study the structure of the UI and we can have a chat afterwards on your observations.

See you :)