Ui updates and outpost hierachy changes
Implemented olgas ui changes and linked it all up. Also given each of the outposts a graphics child so we can control where the art is. The parent of each of the outposts has all thye logic on them. Maybe need to have a discussion of the nav around the main dome???